+++ src/commandsGeneral.js
@@ -4,32 +4,32 @@ import { FormattedMessage } from 'react-intl';
 const commandsGeneral = [
   {
     name: 'expandCollapse',
-    label: (<FormattedMessage id="ui-erm-usage.shortcut.expandCollapse" />),
+    label: (<FormattedMessage id="stripes-components.shortcut.expandOrCollapse" />),
     shortcut: 'spacebar'
   },
   {
     name: 'close',
-    label: (<FormattedMessage id="ui-erm-usage.shortcut.closeModal" />),
+    label: (<FormattedMessage id="stripes-components.shortcut.closeModal" />),
     shortcut: 'esc'
   },
   {
     name: 'copy',
-    label: (<FormattedMessage id="ui-erm-usage.shortcut.copy" />),
+    label: (<FormattedMessage id="stripes-components.shortcut.copy" />),
     shortcut: 'mod+c'
   },
   {
     name: 'cut',
-    label: (<FormattedMessage id="ui-erm-usage.shortcut.cut" />),
+    label: (<FormattedMessage id="stripes-components.shortcut.cut" />),
     shortcut: 'mod+x'
   },
   {
     name: 'paste',
-    label: (<FormattedMessage id="ui-erm-usage.shortcut.paste" />),
+    label: (<FormattedMessage id="stripes-components.shortcut.paste" />),
     shortcut: 'mod+v'
   },
   {
     name: 'find',
-    label: (<FormattedMessage id="ui-erm-usage.shortcut.find" />),
+    label: (<FormattedMessage id="stripes-components.shortcut.find" />),
     shortcut: 'mod+f'
   }
 ];