+++ src/settings/Aggregators/AggregatorForm.js
@@ -196,7 +196,7 @@ class AggregatorForm extends React.Component {
         type="submit"
         disabled={disabled}
       >
-        <FormattedMessage id="ui-erm-usage.udp.form.saveAndClose" />
+        <FormattedMessage id="stripes-components.saveAndClose" />
       </Button>
     );